+++ src/utils/classes/models-interfaces.ts
@@ -1,5 +1,6 @@
+import { BeatsPerBar } from "../../consts/beatsPerBar";
 import { PlayNotes } from "../../consts/playNotes";
-import { PlayFillOn, TimeSignature } from "./patternMaker";
+import { TimeSignature } from "./patternMaker";
 
 export type SubBeatPosition = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7';
 
@@ -11,10 +12,13 @@ export enum MetronomeSounds {
 }
 
 export interface IPatternSettings {
-  playNotes: PlayNotes;
-  playFillOn: PlayFillOn;
-  timeSignature: TimeSignature;
+	playNotes: PlayNotes;
+	timeSignature: TimeSignature;
 }
+
+export type BeatPosition = BeatsPerBar;
+
+export type PlayFillOn = { beat: BeatPosition; subBeat: SubBeatPosition };
 
 // Kept as union of numbers instead of strings as this will never be set by the UI directly.
 // This is only used for calculations and will not be directly part of a metronomeString.
