+++ src/utils/classes/pattern.ts
@@ -1,19 +1,16 @@
 import { IPatternSettings } from "./models-interfaces";
 
 export class Pattern {
-	// metronome sound should not be here?
-	// private metronomeSound: MetronomeSounds = MetronomeSounds.MetronomeSoundOff;
 
 	private defaultPatternSettings: IPatternSettings = {
 		playNotes: 'quarterNotes',
-		playFillOn: { beat: '3', subBeat: '0' }, // can we make these objects into classes of their own or will it be overkill?
-		timeSignature: { beats: '4', division: '4' }, // can we make these objects into classes of their own or will it be overkill?
-	};;
+		timeSignature: { beats: '4', division: '4' },
+	};
 
 	private patternSettings: IPatternSettings; // user provided settings
 
 	constructor(patternSettings?: IPatternSettings) {
-		this.patternSettings = patternSettings ? { ...this.getDefaultPatternSettings(), ...patternSettings} : this.getDefaultPatternSettings(); 
+		this.patternSettings = patternSettings ? { ...this.getDefaultPatternSettings(), ...patternSettings } : this.getDefaultPatternSettings();
 	}
 
 	getDefaultPatternSettings(): IPatternSettings {
@@ -23,7 +20,7 @@ export class Pattern {
 	getPatternSettings(): IPatternSettings {
 		return this.patternSettings;
 	}
-	
+
 	setPatternSettings(patternSettings: IPatternSettings): void {
 		this.patternSettings = patternSettings;
 	}
